rename Stratum to CategoryStratum, part 2/2
