+++ client/lib/CategoryStratum/enableFaceting.js
