+++ client/lib/CategoryStratum/filter.js
