+++ client/lib/CategoryStratum/getBoundingCircle.js
