+++ client/lib/CategoryStratum/getOrigin.js
