+++ client/lib/CategoryStratum/getSpace.js
