+++ client/lib/CategoryStratum/getSubpaths.js
