+++ client/lib/CategoryStratum/index.js
