+++ client/lib/CategoryStratum/layout/index.js
