+++ client/lib/CategoryStratum/layout/normalizeSizes.js
