+++ client/lib/CategoryStratum/layout/rotatePositions.js
