+++ client/lib/CategoryStratum/layout/treepack.js
