+++ client/lib/CategoryStratum/load.js
