+++ client/lib/CategoryStratum/prune.js
