+++ client/lib/CategoryStratum/recomputeBoundingCircle.js
