+++ client/lib/CategoryStratum/refreshNodeSizes.js
