+++ client/lib/CategoryStratum/remove.js
