+++ client/lib/CategoryStratum/render.js
