+++ client/lib/CategoryStratum/renderContextLabel.js
