+++ client/lib/CategoryStratum/revealLabels.js
