+++ client/lib/CategoryStratum/stratum.css
